py-scripts/requirements.txt
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.1
idna==3.10
joblib==1.5.1
jsonpickle==4.1.1
kiwisolver==1.4.8
matplotlib==3.10.3
more-itertools==10.7.0
music21==9.7.0
mutagen==1.47.0
numpy==1.26.4
packaging==25.0
pillow==11.2.1
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
tqdm==4.67.1
urllib3==2.4.0
webcolors==24.11.1
